vmm: nested_vt: refine code layout
split nested_vt.c into several c files.
